server/requirements.txt
Flask==2.1.0
flask-cors==3.0.10
gunicorn==20.1.0
numpy==1.21.0
openai==0.10.2
python-dotenv==0.17.1
requests==2.25.1
textblob==0.15.3
